+++ src/tournament_index/src/cycle.rs
@@ -24,7 +24,7 @@ async fn request_cycles() -> Result<(), TournamentIndexError> {
         ));
     }
 
-    transfer_cycles(CYCLES_TOP_UP_AMOUNT as u128, caller).await
+    transfer_cycles(CYCLES_TOP_UP_AMOUNT, caller).await
 }
 
 async fn transfer_cycles(
